+++ korge-kotlin-compiler/src/main/kotlin/korlibs/korge/kotlincompiler/KorgeKotlinCompiler.kt
@@ -2,60 +2,54 @@
 
 package korlibs.korge.kotlincompiler
 
+import korlibs.korge.kotlincompiler.maven.*
 import org.jetbrains.kotlin.buildtools.api.*
 import org.jetbrains.kotlin.buildtools.api.jvm.*
 import org.jetbrains.kotlin.daemon.common.*
 import org.jetbrains.kotlin.utils.addToStdlib.*
-import org.w3c.dom.*
 import java.io.*
-import java.net.*
 import java.security.*
 import java.util.*
-import javax.xml.*
-import javax.xml.parsers.*
-import kotlin.collections.ArrayDeque
 
 // https://github.com/JetBrains/kotlin/tree/master/compiler/build-tools/kotlin-build-tools-api
 // https://github.com/JetBrains/kotlin/blob/bc1ddd8205f6107c7aec87a9fb3bd7713e68902d/compiler/build-tools/kotlin-build-tools-api-tests/src/main/kotlin/compilation/model/JvmModule.kt
 class KorgeKotlinCompiler {
-    data class Module(
-        val path: File,
-        val moduleDeps: Set<Module> = setOf(),
-        val libs: Set<File> = setOf(),
-        val main: String = "MainKt",
-    ) {
-        val allModuleDeps: Set<Module> by lazy {
-            (moduleDeps.flatMap { it.allModuleDeps + it } + this).toSet()
-        }
-        val allTransitiveLibs by lazy { allModuleDeps.flatMap { it.libs }.toSet() }
-
-        val buildDir = File(path, ".korge")
-        val classesDir = File(buildDir, "classes")
-        val srcDirs by lazy {
-            val dirs = arrayListOf<File>()
-            if (File(path, "src/commonMain/kotlin").isDirectory) {
-                dirs += File(path, "src/commonMain/kotlin")
-                dirs += File(path, "src/jvmMain/kotlin")
-            } else {
-                dirs += File(path, "src")
-                dirs += File(path, "src@jvm")
-            }
-            dirs
-        }
-        val resourceDirs by lazy {
-            val dirs = arrayListOf<File>()
-            if (File(path, "src/commonMain/kotlin").isDirectory) {
-                dirs += File(path, "src/commonMain/resources")
-                dirs += File(path, "src/jvmMain/resources")
-            } else {
-                dirs += File(path, "resources")
-                dirs += File(path, "resources@jvm")
-            }
-            dirs
-        }
-    }
 
     companion object {
+
+        @JvmStatic
+        fun main(args: Array<String>) {
+            println(javaExecutablePath)
+
+            val libs = filesForMaven(
+                MavenArtifact("org.jetbrains.kotlin", "kotlin-stdlib", "2.0.0"),
+                MavenArtifact("com.soywiz.korge", "korge-jvm", "999.0.0.999")
+            )
+
+            val mod1 = Module(
+                path = File("C:\\Users\\soywiz\\projects\\korge-snake\\modules\\korma-tile-matching"),
+                libs = libs
+            )
+            val snakeModule = Module(
+                path = File("C:\\Users\\soywiz\\projects\\korge-snake"),
+                moduleDeps = setOf(mod1),
+                libs = libs
+            )
+
+            compileAndRun(snakeModule, mapOf("KORGE_HEADLESS" to "true", "KORGE_IPC" to "C:\\Users\\soywiz\\AppData\\Local\\Temp\\/KORGE_IPC-304208"))
+        }
+
+        fun compileAndRun(module: Module, envs: Map<String, String> = mapOf()) {
+            compileAllModules(module)
+            runModule(module, envs)
+        }
+
+        fun compileAllModules(module: Module) {
+            module.allModuleDeps.forEach {
+                compileModule(it)
+            }
+        }
+
         val javaExecutablePath by lazy {
             ProcessHandle.current()
                 .info()
@@ -91,9 +85,9 @@ class KorgeKotlinCompiler {
             //compiler.runJvm()
         }
 
-        fun runModule(module: Module) {
+        fun runModule(module: Module, envs: Map<String, String> = mapOf()) {
             val allClasspaths: Set<File> = module.allModuleDeps.flatMap { setOf(it.classesDir) + it.resourceDirs + it.libs }.toSet()
-            runJvm(module.main, allClasspaths)
+            runJvm(module.main, allClasspaths, envs)
         }
 
         fun runJvm(main: String, classPaths: Collection<File>, envs: Map<String, String> = mapOf()): Int {
@@ -103,88 +97,50 @@ class KorgeKotlinCompiler {
                 .inheritIO()
                 .also { it.environment().putAll(envs) }
                 .start()
+                .also { proc -> Runtime.getRuntime().addShutdownHook(Thread { proc.destroy(); proc.destroyForcibly() }) }
                 .waitFor()
-        }
-
-
-        @JvmStatic
-        fun main(args: Array<String>) {
-            println(javaExecutablePath)
-
-            val libs = filesForMaven(
-                MavenArtifact("org.jetbrains.kotlin", "kotlin-stdlib", "2.0.0"),
-                MavenArtifact("com.soywiz.korge", "korge-jvm", "999.0.0.999")
-            )
-
-            val mod1 = Module(
-                path = File("C:\\Users\\soywiz\\projects\\korge-snake\\modules\\korma-tile-matching"),
-                libs = libs
-            )
-            val mod2 = Module(
-                path = File("C:\\Users\\soywiz\\projects\\korge-snake"),
-                moduleDeps = setOf(mod1),
-                libs = libs
-            )
-
-            compileModule(mod1)
-            compileModule(mod2)
-            runModule(mod2)
-
-            /*
-            return
-
-            run {
-                val compiler = KorgeKotlinCompiler()
-                compiler.buildDirectory = File("C:\\temp\\.kotlin")
-                compiler.rootDir = File("C:\\temp")
-                compiler.sourceDirs = setOf(
-                    File("C:\\temp\\1"),
-                    File("C:\\temp\\1-common"),
-                )
-                compiler.libs = filesForMaven(
-                    MavenArtifact("org.jetbrains.kotlin", "kotlin-stdlib", "2.0.0"),
-                    MavenArtifact("com.soywiz.korge", "korge-jvm", "999.0.0.999")
-                )
-
-                val (time, result) = measureTimeMillisWithResult {
-                    compiler.compileJvm()
-                    //println(compiler.compileJvm(forceRecompilation = true))
-                }
-                println("$result: $time ms")
-                compiler.runJvm()
-            }
-
-            return
-            repeat(10) {
-                run {
-
-                }
-
-                val compiler = KorgeKotlinCompiler()
-                compiler.buildDirectory = File("C:\\Users\\soywiz\\projects\\korge-snake\\.kotlin")
-                compiler.rootDir = File("C:\\Users\\soywiz\\projects\\korge-snake")
-                compiler.sourceDirs = setOf(
-                    File("C:\\Users\\soywiz\\projects\\korge-snake\\src"),
-                    File("C:\\Users\\soywiz\\projects\\korge-snake\\modules\\korma-tile-matching\\src\\commonMain\\kotlin"),
-                )
-                compiler.libs = filesForMaven(
-                    MavenArtifact("org.jetbrains.kotlin", "kotlin-stdlib", "2.0.0"),
-                    MavenArtifact("com.soywiz.korge", "korge-jvm", "999.0.0.999")
-                )
-
-                repeat(4) { NN ->
-                    println(measureTimeMillis {
-                        println(compiler.compileJvm(forceRecompilation = NN == 0))
-                        //println(compiler.compileJvm(forceRecompilation = true))
-                    })
-                }
-            }
-            */
         }
 
         fun filesForMaven(vararg artifacts: MavenArtifact): Set<File> = artifacts.flatMap { filesForMaven(it) }.toSet()
         fun filesForMaven(artifacts: List<MavenArtifact>): Set<File> = artifacts.flatMap { filesForMaven(it) }.toSet()
         fun filesForMaven(artifact: MavenArtifact): Set<File> = MavenTools.getMavenArtifacts(artifact)
+    }
+
+    data class Module(
+        val path: File,
+        val moduleDeps: Set<Module> = setOf(),
+        val libs: Set<File> = setOf(),
+        val main: String = "MainKt",
+    ) {
+        val allModuleDeps: Set<Module> by lazy {
+            (moduleDeps.flatMap { it.allModuleDeps + it } + this).toSet()
+        }
+        val allTransitiveLibs by lazy { allModuleDeps.flatMap { it.libs }.toSet() }
+
+        val buildDir = File(path, ".korge")
+        val classesDir = File(buildDir, "classes")
+        val srcDirs by lazy {
+            val dirs = arrayListOf<File>()
+            if (File(path, "src/commonMain/kotlin").isDirectory) {
+                dirs += File(path, "src/commonMain/kotlin")
+                dirs += File(path, "src/jvmMain/kotlin")
+            } else {
+                dirs += File(path, "src")
+                dirs += File(path, "src@jvm")
+            }
+            dirs
+        }
+        val resourceDirs by lazy {
+            val dirs = arrayListOf<File>()
+            if (File(path, "src/commonMain/kotlin").isDirectory) {
+                dirs += File(path, "src/commonMain/resources")
+                dirs += File(path, "src/jvmMain/resources")
+            } else {
+                dirs += File(path, "resources")
+                dirs += File(path, "resources@jvm")
+            }
+            dirs
+        }
     }
 
     var rootDir: File = File("/temp")
@@ -340,70 +296,4 @@ class KorgeKotlinCompiler {
     }
 }
 
-data class MavenArtifact(val group: String, val name: String, val version: String, val classifier: String? = null, val extension: String = "jar") {
-    val groupSeparator by lazy { group.replace(".", "/") }
-    val localPath by lazy { "$groupSeparator/$name/$version/$name-$version.$extension" }
-}
-data class MavenDependency(val artifact: MavenArtifact, val scope: String)
 
-class Pom(
-    val packaging: String? = null,
-    val deps: List<MavenDependency> = emptyList(),
-) {
-    companion object {
-        fun parse(file: File): Pom = parse(file.readText())
-        fun parse(text: String): Pom {
-            val db = DocumentBuilderFactory.newInstance().also { it.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true) }.newDocumentBuilder()
-            val doc = db.parse(text.byteInputStream())
-            val out = arrayListOf<MavenDependency>()
-            val node = doc.getElementsByTagName("packaging").toList().firstOrNull()
-            for (e in doc.getElementsByTagName("dependency").toList()) {
-                val groupId = e.findChildByTagName("groupId").firstOrNull()?.textContent?.trim() ?: error("Missing groupId")
-                val artifactId = e.findChildByTagName("artifactId").firstOrNull()?.textContent?.trim() ?: error("Missing artifactId")
-                val scope = e.findChildByTagName("scope").firstOrNull()?.textContent?.trim()
-                if (scope == "test" || scope == null) continue
-                val version = e.findChildByTagName("version").firstOrNull()?.textContent?.trim() ?: error("Missing version for $groupId:$artifactId in $text")
-                if (version.contains("\$")) continue
-                out += MavenDependency(MavenArtifact(groupId, artifactId, version), scope ?: "compile")
-                //println("DEP: $groupId:$artifactId:$version  :: $scope")
-            }
-            return Pom(packaging = node?.textContent, deps = out)
-        }
-
-        private fun NodeList.toList(): List<Node> = (0 until length).map { item(it) }
-        private fun Node.findChildByTagName(tagName: String): List<Node> = childNodes.toList().filter { it.nodeName.equals(tagName, ignoreCase = true) }
-    }
-}
-
-object MavenTools {
-    fun getMavenArtifacts(artifact: MavenArtifact, explored: MutableSet<MavenArtifact> = mutableSetOf()): Set<File> {
-        val explore = ArrayDeque<MavenArtifact>()
-        explore += artifact
-        val out = mutableSetOf<MavenArtifact>()
-        while (explore.isNotEmpty()) {
-            val artifact = explore.removeFirst()
-            if (artifact in explored) continue
-            explored += artifact
-            val pom = Pom.parse(getSingleMavenArtifact(artifact.copy(extension = "pom")))
-            if (pom.packaging == null || pom.packaging == "jar") {
-                out += artifact
-            }
-            for (dep in pom.deps) {
-                explore += dep.artifact
-            }
-        }
-        return out.map { getSingleMavenArtifact(it) }.toSet()
-    }
-
-    fun getSingleMavenArtifact(artifact: MavenArtifact): File {
-        val file = File(System.getProperty("user.home"), ".m2/repository/${artifact.localPath}")
-        if (!file.exists()) {
-            file.parentFile.mkdirs()
-            val url = URL("https://repo1.maven.org/maven2/${artifact.localPath}")
-            println("Downloading $url")
-            file.writeBytes(url.readBytes())
-        }
-        return file
-    }
-
-}
