+++ korge-kotlin-compiler/src/main/kotlin/korlibs/korge/kotlincompiler/maven/MavenDependency.kt
@@ -1,0 +1,3 @@
+package korlibs.korge.kotlincompiler.maven
+
+data class MavenDependency(val artifact: MavenArtifact, val scope: String)
